Implement endpoint for psuedo/unsecure authentication
`GET` requests to the `/authenticate` endpoint are now handled. The `password` and
`username` parameters are required and compared against those stored in the database.
If the user is found and the password query parameter matches the password in the
database, a 200 status code is sent. Otherwise, an error status code is sent.