Implement updateFavorites
The `updateFavorites` function requests the list of books associated with the current user
from the server and updates the `favorites` in state with the retrieved array. The code
performing the same functionality in the `useEffect` hook has been replaced with a call of
`updateFavorites`, and `updateFavorites` has been passed down the FavoriteList for use in
the `AddBookForm` component.