Replace Home with Overview
The Home component has been renamed to "Overview" to prevent confusion with the landing page
component. It also now retrieves the value for the user state variable from the "user"
parameter in the URL so different users can log in. Previously the initial value for the
user state variable was hard-coded.