Fix notes without a page value not being able to be added
Since the state variable for handling the value of the page field on the note creation
form was set to an empty string initially, if it was left unedited (as it is an optional
field) it would be sent to the server as a string. As the database does not accept a
string as a valid value for the page column, and error occurred. This has been fixed
by setting the page value to null before making the request to the server.