Fix NoteList component re-rendering infinitely
The NoteList component was updating state continuously as well as sending requests to the
server to get the notes data constantly. This was due to a useEffect hook being run
constantly. This has been fixed by setting the hook to only run the function on the
initial render. Consequently, state has been raised to the parent component as well as
the updateNotes function so that sibling components could access it, now that the updateNotes
function is necessary to update state when needed.